pyproject.toml
[project]
name = "dipm"
version = "0.0.1"
description = "Deep Interatomic Potential Models in JAX"
license = "Apache 2.0"
authors = [
    {name = "Cao Bohan", email = "caobohan@mail.nankai.edu.cn"},
    {name = "InstaDeep", email = "c.brunken@instadeep.com"},
]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "numpy (>=1.26.0,<2.0.0)",
    "jax (>=0.4.33)",
    "orbax-checkpoint (==0.5.2)",
    "omegaconf (>=2.3.0)",
    "e3nn-jax (>=0.20.7)",
    "flax (>=0.9.0)",
    "jraph",
    "jax-md (>=0.2.8)",
    "orbax (>=0.1.9)",
    "matscipy (==1.0.0)",
    "tdqm (>=0.0.1)",
    "pydantic (>=2.10.2)",
    "tqdm-loggable (>=0.2)",
    "h5py (>=3.12.1)",
    "safetensors (>=0.5.0)",
]

[project.optional-dependencies]
# TODO: Use cue to accelerate e3nn
cue = [
    "cuequivariance (>=0.6.0)",
    "cuequivariance-jax (>=0.6.0)",
]

[tool.poetry.urls]
"Homepage" = "https://github.com/bhcao/deep-interatomic-potential-models"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "24.2.0"
pre-commit = "^4.0.1"
sphinx = "^8.1.3"
furo = "^2024.8.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.coverage.report]
include_namespace_packages = true
